third_party/blink/web_tests/platform/generic/printing/page-count-percentage-height-expected.txt

Test to make sure 2nd page is printed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1" is 0
PASS: page number of "page2" is 1
PASS: number of pages is 2
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
